chore: move attributes to a separate file
