+++ src/blocks/blocks/pricing-table/group/attributes.js
@@ -1,0 +1,14 @@
+const attributes = {
+	hasMoneyBackSection: {
+		type: 'boolean',
+		default: false
+	},
+	title: {
+		type: 'string'
+	},
+	text: {
+		type: 'string'
+	}
+};
+
+export default attributes;
